remove LiveWidget component and integrate live status polling into LogoWidget
